% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringManipulation.R
\name{charSeq}
\alias{charSeq}
\title{Generate a squence of string with prefix and suffix}
\usage{
charSeq(prefix = "", from, to, by = 1, suffix = "")
}
\arguments{
\item{prefix}{prefix of the sequence}

\item{from, to}{the starting and (maximal) end values of the sequence}

\item{by}{number: increment of the sequence.}

\item{suffix}{suffix of the sequence}
}
\value{
character vector
}
\description{
This function generate a sequence of string like: \code{prefix001, prefix002,
prefix003 ... prefix999}
}
\examples{
charSeq(prefix = "id", from = 1, to = 10)
charSeq(prefix = "H-", from = 10, to = 0, by = -1, suffix = "s")
}
