% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages2install.R
\name{.findPackages}
\alias{.findPackages}
\title{Exctract the packages loaded in a R file}
\usage{
.findPackages(file)
}
\arguments{
\item{file}{R file}
}
\value{
character vector
}
\description{
Exctract the packages loaded in a R file
}
