% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages2install.R
\name{packages2Install}
\alias{packages2Install}
\title{Find all R packages to install}
\usage{
packages2Install(path = getwd(), file = NA, quote = "\\"")
}
\arguments{
\item{path}{path of the R project (current working directory by default)}

\item{file}{(optional) file name of the output file}

\item{quote}{(optional) quote type, either " or '  (default: " )}
}
\description{
This function will check all ".R" file in a given repository and find all
refferences to loaded packages through \code{library(pkg)} or
\code{require(pkg)}. The function will show a R line to install these
packages: \code{install.packages(c("pkg1", "pkg2", ...)}
}
\details{
NOTE : This work for packages hosted on CRAN, not on Bioconductor. More
generally it do not work for package than can't be installed by :
\code{install.packages("pkg")} do not work
}
\examples{
packages2Install()
}
