% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_plot.R
\name{plotdnorm}
\alias{plotdnorm}
\title{plotdnorm}
\usage{
plotdnorm(
  mean = 0,
  sd = 1,
  from = mean - 5 * sd,
  to = mean + 5 * sd,
  by = 0.01,
  log = FALSE,
  ...
)
}
\arguments{
\item{mean}{mean}

\item{sd}{standard deviations}

\item{from, to}{x limits}

\item{by}{number: increment of the sequence}

\item{log}{logical; if TRUE, probabilities p are given as log(p)
(see \code{\link[stats]{dnorm}})}

\item{...}{Other rguments to be passed to \code{plot_ly} (or \code{plot})
functions}
}
\value{
An interactive graph if \pkg{plotly} package is available or a basic R plot.
}
\description{
This function plot the density function of a normal distribution.
}
\author{
Julien Diot
}
